Added test case for the aye_recognition (#42)
