+++ tox.ini
@@ -7,6 +7,6 @@ commands = {envpython} setup.py test
          
 [testenv:style]
 deps = -rrequirements.txt
-whitelist_externals= bash
+whitelist_externals = bash
 commands = bash -c \'find . -iname "*.py" -not -regex "\(./venv/.*\)\|\(./.tox/.*\)\|\(./.eggs/.*\)" | xargs flake8\'   
            bash -c \'find . -iname "*.py" -not -regex "\(./venv/.*\)\|\(./.tox/.*\)\|\(./.eggs/.*\)" | xargs pylint\'
